requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.11.0
django-jwt-auth==0.0.2
django-rest-authtoken==2.1.4
djangorestframework==3.13.1
mysqlclient==2.1.0
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
